Directed replacements
Add support to direct replacements at users.  This allows one user to
correct another.  A message such as "wronguser: s/wrong/correct/g" will
perform the search and replace on wronguser's last message.

Currently the nick and replacement command be be delimited by one of
[;, ], followed by whitespace.

The user's last message is updated with the correction, so if they
decide to correct themselves, it will have to be based on the updated
message.